requirements.txt
-i https://pypi.org/simple
contourpy==1.2.0 ; python_version >= '3.9'
cycler==0.12.1 ; python_version >= '3.8'
exceptiongroup==1.1.3 ; python_version < '3.11'
fonttools==4.44.0 ; python_version >= '3.8'
iniconfig==2.0.0 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2 ; python_version >= '3.7'
pandas==2.1.2
pillow==10.1.0 ; python_version >= '3.8'
pluggy==1.3.0 ; python_version >= '3.8'
psycopg2==2.9.9
pyparsing==3.1.1 ; python_full_version >= '3.6.8'
pytest==7.4.3
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.8.0 ; python_version >= '3.8'
tzdata==2023.3 ; python_version >= '2'
